Rewrite pfctl_test in C
Even with the latest optimizations to atf-sh this test still takes a few
seconds to startup in QEMU. Re-writing it in C reduces the runtime for a
single test from about 3 seconds to .5 seconds. Since there are ~80 tests,
this adds up to about 3-4 minutes. This may not seem like a big speedup,
but before the recent optimizations to atf-sh each test took almost 100
seconds on QEMU RISC-V instead of 3.