Massively speed up mkimg_test
It turns out that the real problem with the mkimg tests isn't mkimg itself
but rather the use of jot and hexdump which for some reason are **much**
slower as purecap binaries than non-CHERI ones. Interestingly this does
not make much of a time difference when using upstream FreeBSD on CHERI
QEMU, so it must be related to the CheriBSD kernel.